研修成果物発表時

+++ app/Http/Controllers/BooksController.php
@@ -20,13 +20,26 @@ class BooksController extends Controller
         if(\Auth::check()){
             $keyword=$request->input('word');
             $query=Book::query();
+            //検索結果
+            $count=0;
+            //ほんの冊数
+            $totalNumber=0;
             if(!empty($keyword)){
                 $query->where('book_display','LIKE',"%{$keyword}%");
+                $books=$query->paginate(5);
+                $count=$query->count();
+                foreach($books as $book){
+                    $totalNumber += $book->book_number;
+                }
             }else{
                 //本の一覧を取得
-                $books=Book::all();                
+                $books=Book::orderBy('id','desc')->paginate(5);
+                //ほんのカウント用
+                $book_count=Book::all();
+                foreach($book_count as $book){
+                $totalNumber += $book->book_number;
+                }
             }
-            $books=$query->get();
 
             //ユーザー情報
             $user=\Auth::user();
@@ -34,7 +47,9 @@ class BooksController extends Controller
             return view('books.index',[
                 'user'=>$user,
                 'keyword'=>$keyword,
-                'books'=>$books,            
+                'books'=>$books,
+                'count'=>$count,
+                'number'=>$totalNumber,
             ]);
         }
         return view('dashboard');
@@ -50,6 +65,7 @@ class BooksController extends Controller
         if(\Auth::check()){
             $book = new Book;
             $user=\Auth::user();
+
             //登録ビューを表示
             return view('books.create',[
                'book' => $book,
@@ -74,6 +90,10 @@ class BooksController extends Controller
                 'book_title'=>$request->book_title,
                 'book_display'=>$request->book_display,
                 'book_number'=>$request->book_number,
+                'book_isbn'=>$request->book_isbn,
+                'book_author'=>$request->book_author,
+                'book_publisher'=>$request->book_publisher,
+                'book_coverURL'=>$request->book_coverURL,
             ]);
         }
         //トップページへリダイレクト
@@ -91,15 +111,16 @@ class BooksController extends Controller
         //idの値でコミックを検索して取得
         $book = Book::findOrFail($id);
         if(\Auth::check()){
-            $user=\Auth::user();
-            //詳細ビューで表示
-            return view('books.show',[
-                'book' =>$book,
-                'user' => $user,
-            ]);
-        }else{
-            return redirect('/');
+            if(\Auth::id()===$book->user_id){
+                $user=\Auth::user();
+                //詳細ビューで表示
+                return view('books.show',[
+                    'book' =>$book,
+                    'user' => $user,
+                ]);
+            }
         }
+        return redirect('/');
     }
 
     /**
@@ -112,17 +133,17 @@ class BooksController extends Controller
     {
         //idの値でコミックを検索して取得
         $book = Book::findOrFail($id);
-        $user=\Auth::user();
-        if(\Auth::id()===$book->user_id){
-            //表示
-            return view('books.edit',[
-                'book' => $book,
-                'user' => $user,
-            ]);            
-        }else{
-            return redirect('/');
+        if(\Auth::check()){
+            $user=\Auth::user();
+            if(\Auth::id()===$book->user_id){
+                //表示
+                return view('books.edit',[
+                    'book' => $book,
+                    'user' => $user,
+                ]);            
+            }
         }
-
+        return redirect('/');
     }
 
     /**
@@ -132,53 +153,44 @@ class BooksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    //タイトル更新
-    public function title_update(Request $request,$id){
-        //idの値でコミックを検索して取得
+    //本の情報更新
+    public function info_update(Request $request,$id){
         $book = Book::findOrFail($id);
-        if(\Auth::id()===$book->user_id){
-            //コミック情報を更新
-            $book->book_title = $request->book_title;
-            $book->save();
-        }
-        return redirect('/');
-    }
-    //表示名更新
-    public function display_update(Request $request,$id){
-        //idの値でコミックを検索して取得
-        $book = Book::findOrFail($id);
-        if(\Auth::id()===$book->user_id){
-            //コミック情報を更新
-            $book->book_display = $request->book_display;
-            $book->save();
-        }
-        return redirect('/');
-    }
-    //冊数更新
-    public function number_update(Request $request,$id){
-        //idの値でコミックを検索して取得
-        $book = Book::findOrFail($id);
-        if(\Auth::id()===$book->user_id){
-            //コミック情報を更新
-            $book->book_number = $request->book_number;
-            $book->save();
+        if(\Auth::check()){
+            if(\Auth::id()===$book->user_id){
+                //コミック情報を更新
+                $book->book_title = $request->book_title;
+                $book->book_isbn = $request->book_isbn;
+                $book->book_author = $request->book_author;
+                $book->book_publisher = $request->book_publisher;
+                $book->book_coverURL = $request->book_coverURL;
+                $book->book_display = $request->book_display;
+                $book->book_number = $request->book_number;
+                $book->save();
+            }
         }
         return redirect('/');
     }
     //冊数の変更(ボタン)
     public function number_fluctuation_plus(Request $request,$id){
         $book=Book::findOrFail($id);
-        if(\Auth::id()===$book->user_id){
-            $book->book_number +=1;
-            $book->save();
+        if(\Auth::check()){
+            if(\Auth::id()===$book->user_id){
+                $book->book_number +=1;
+                $book->save();
+            }
         }
         return back();
     }
     public function number_fluctuation_minus(Request $request,$id){
         $book=Book::findOrFail($id);
-        if(\Auth::id()===$book->user_id){
-            $book->book_number -=1;
-            $book->save();
+        if(\Auth::check()){
+            if(\Auth::id()===$book->user_id){
+                if(0<$book->book_number){
+                    $book->book_number -=1;
+                    $book->save();                
+                }
+            }
         }
         return back();
     }
@@ -192,9 +204,10 @@ class BooksController extends Controller
     {
         //idの値でコミックを検索して取得
         $book = Book::findOrFail($id);
-        //コミックを削除
-        $book->delete();
-        
+        if(\Auth::check()){
+            //コミックを削除
+            $book->delete();
+        }
         //トップページへリダイレクト
         return redirect('/');
     }
